make ray client separate
