+++ mars/deploy/oscar/tests/test_ray.py
@@ -16,11 +16,6 @@ import asyncio
 import copy
 import operator
 import os
-import subprocess
-import sys
-import tempfile
-import threading
-import time
 from functools import reduce
 
 import numpy as np
@@ -237,59 +232,6 @@ def new_ray_session_test():
 
 
 @require_ray
-def test_ray_client():
-    server_code = """import time
-import ray.util.client.server.server as ray_client_server
-
-server = ray_client_server.init_and_serve("{address}", num_cpus=20)
-print("OK", flush=True)
-while True:
-    time.sleep(1)
-"""
-
-    address = "127.0.0.1:50051"
-
-    with tempfile.NamedTemporaryFile(mode="w", suffix=".py") as f:
-        f.write(server_code.format(address=address))
-        f.flush()
-
-        proc = subprocess.Popen([sys.executable, "-u", f.name], stdout=subprocess.PIPE)
-
-        try:
-
-            def _check_ready(expect_exit=False):
-                while True:
-                    line = proc.stdout.readline()
-                    if proc.returncode is not None:
-                        if expect_exit:
-                            break
-                        raise Exception(
-                            f"Failed to start ray server at {address}, "
-                            f"the return code is {proc.returncode}."
-                        )
-                    if b"OK" in line:
-                        break
-
-            # Avoid ray.init timeout.
-            _check_ready()
-
-            # Avoid blocking the subprocess when the stdout pipe is full.
-            t = threading.Thread(target=_check_ready, args=(True,))
-            t.start()
-
-            ray.init(f"ray://{address}")
-            ray._inside_client_test = True
-            try:
-                new_ray_session_test()
-            finally:
-                ray._inside_client_test = False
-                ray.shutdown()
-        finally:
-            proc.kill()
-            proc.wait()
-
-
-@require_ray
 @pytest.mark.parametrize(
     "test_option",
     [
@@ -429,8 +371,8 @@ async def test_load_third_party_modules_from_config(
         config=CONFIG_THIRD_PARTY_MODULES_TEST_FILE,
     )
     async with client:
-        # 1 supervisor, 2 worker main pools, 4 worker sub pools.
-        assert len(get_output_filenames()) == 7
+        # 1 supervisor, 1 worker main pools, 1 worker sub pools.
+        assert len(get_output_filenames()) == 3
 
 
 @require_ray
